% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cnv.R
\name{getCNVRatioFromControl}
\alias{getCNVRatioFromControl}
\title{Get CNV ratios from a control sample. Control should be binned and processed similarly to treatment sample}
\usage{
getCNVRatioFromControl(counts_df, genome, meta_bin_size = 50, n_windows = 10)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, counts}

\item{genome}{genome name For example, hg19, hg38, mm10}

\item{meta_bin_size}{param passed to pbsR:::testCNVPresence}

\item{n_windows}{param passed to pbsR:::testCNVPresence}
}
\description{
Get CNV ratios from a control sample. Control should be binned and processed similarly to treatment sample
}
