% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cnv.R
\name{getCNVWithCNAnorm}
\alias{getCNVWithCNAnorm}
\title{Calculate CNV ratios}
\usage{
getCNVWithCNAnorm(counts_df, reference_k = 15.7, reference_beta = 4.81)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, and counts.}

\item{reference_k}{TODO: add DOC}

\item{reference_beta}{TODO: add DOC}
}
\description{
Calculate CNV ratios
}
