% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{getDistributionParametersWithOptim}
\alias{getDistributionParametersWithOptim}
\title{Get Distribution Parameters}
\usage{
getDistributionParametersWithOptim(
  working_df,
  theta = 0.5,
  lambda_range = seq(0.5, 1, by = 0.1),
  param_search_length = 3,
  fix_weight = TRUE,
  weight_value = 500,
  max_dens = 20
)
}
\arguments{
\item{working_df}{dataframe with one column counts}

\item{theta}{estimate of noise (default 0.5)}

\item{lambda_range}{range of lambda values to test}

\item{param_search_length}{number of parameters to generate, per parameter}

\item{fix_weight}{if true, fix_weight value to be 0.99 of max(working_df$counts), else set to weight_value=500}

\item{weight_value}{fix tail end of distribution}

\item{max_dens}{x-axis cut-off on density plot}
}
\description{
This function takes in a dataframe with atleast one column named counts (typically counts in genomics bins).
It will retrieve some initial distribution parameters based on the count data, and fine-tune the parameters by fitting
parameters to the pgamma.null distribution until convergence (low CvM statistic).
}
