% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cnv.R
\name{getGCContent}
\alias{getGCContent}
\title{Given set of genomic coordinates in a dataframe (chr, start, end), get the GC content using BSGenome libraries}
\usage{
getGCContent(bed_df, genome, bin_size = 0)
}
\arguments{
\item{bed_df}{dataframe with columns chr, start, end.}

\item{genome}{genome name For example, hg19, hg38, mm10}

\item{bin_size}{size of genome wide bins. If not user defined, will calculate from bed_df}
}
\description{
Given set of genomic coordinates in a dataframe (chr, start, end), get the GC content using BSGenome libraries
}
