% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{getInitialDistributionParameters}
\alias{getInitialDistributionParameters}
\title{Get Initial Distribution Parameters}
\usage{
getInitialDistributionParameters(
  working_df,
  theta = 0.5,
  lambda_range = seq(from = 0.4, to = 0.8, by = 0.05),
  fix_weight = TRUE,
  weight_value = 500,
  use_log = FALSE,
  param_search_length = 100,
  max_range_multiple = 300,
  plot_data = FALSE
)
}
\arguments{
\item{working_df}{dataframe with one column counts}

\item{theta}{estimate of noise (default 0.5)}

\item{lambda_range}{range of lambda values to test}

\item{fix_weight}{if true, fix_weight value to be 0.99 of max(working_df$counts), else set to weight_value=500}

\item{weight_value}{fix tail end of distribution}

\item{param_search_length}{number of parameters to generate, per parameter}

\item{max_dens}{x-axis cut-off on density plot}
}
\description{
This function takes in a dataframe with atleast one column named counts (typically counts in genomics bins).
It will generate some initial values of lambda, beta and k based on the minimum CvM statistic returned by testing a
coarse grid of parameters.
}
