% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_counts.R
\name{rescaleMappability}
\alias{rescaleMappability}
\title{Rescale counts using mappability scores}
\usage{
rescaleMappability(counts_df, map_df, map_threshold = 0.5)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, and counts. Generally, the output of pbsR::getBinnedCounts.}

\item{map_df}{dataframe with columns chr, start, end, and mappability_score. Look at get("hg19_5000_map_100", asNamespace('pbsR')) to see an example}
}
\description{
Rescale counts using mappability scores
}
