% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_counts.R
\name{RescaleXY}
\alias{RescaleXY}
\title{Rescale mappability scores in chrX and chrY}
\usage{
RescaleXY(counts_df)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, and counts. Generally, the input to this function is the output of pbsR::RescaleMappability}
}
\description{
Rescale mappability scores in chrX and chrY
}
