% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cnv.R
\name{testCNVPresence}
\alias{testCNVPresence}
\title{Check for CNVs based on the presence of a bimodal distribution in genome-wide binned counts.}
\usage{
testCNVPresence(counts_df, meta_bin_size = 50, n_windows = 10, bin_size = 0)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, and counts. Generally, the input is the output of pbsR::getMappabilityScore}

\item{meta_bin_size}{TODO: add DOC}

\item{n_windows}{TODO: add DOC}
}
\description{
Check for CNVs based on the presence of a bimodal distribution in genome-wide binned counts.
}
