% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_counts.R
\name{getBinnedCounts}
\alias{getBinnedCounts}
\title{Bin genome given chromosome sizes, and count number of reads within each bin}
\usage{
getBinnedCounts(bam_file, chrom_sizes_file, bin_size, paired_end, threads = 4)
}
\arguments{
\item{bam_file}{filepath to BAM file. BAM file must be indexed}

\item{chrom_sizes_file}{filepath to chromosome sizes. Can get from UCSC genome browser.}

\item{bin_size}{size of bin to split genome into non-overlapping windows}

\item{paired_end}{true/false if dataset is paired/single read}

\item{threads}{number of threads to run featureCounts}
}
\description{
This function takes in filepath that specifies chromosome lengths, bins the genome according to bin_size and
gets the count of reads (single-end) or fragments (paired-end) using featureCounts
}
