% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cnv.R
\name{getCNVRatioFromControl}
\alias{getCNVRatioFromControl}
\title{Get CNV ratios from a control sample. Control should be binned and processed similarly to treatment sample}
\usage{
getCNVRatioFromControl(counts_df, meta_bin_size = 50, n_windows = 10)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, counts and map_rescaled_counts.}

\item{meta_bin_size}{param passed to pbsR::TestCNVPresence}

\item{n_windows}{param passed to pbsR::TestCNVPresence}
}
\description{
Get CNV ratios from a control sample. Control should be binned and processed similarly to treatment sample
}
