% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_counts.R
\name{getMappabilityScore}
\alias{getMappabilityScore}
\title{Get mappability score for each bin in binned genome, and scale counts according to mappability score}
\usage{
getMappabilityScore(
  bam_file,
  counts_df,
  bin_size = 0,
  genome,
  paired_end,
  map_threshold = 0.5
)
}
\arguments{
\item{bam_file}{filepath to BAM file. BAM file must be indexed}

\item{counts_df}{dataframe with columns chr, start, end, and counts. Generally, the input to this function is the output of pbsR::getBinnedCounts}

\item{bin_size}{size of bin to split genome into non-overlapping windows. If not supplied, will extract from counts_df}

\item{paired_end}{true/false if dataset is paired/single read}

\item{map_threshold}{bins with mappability score lower than this threshold will be filtered out}
}
\description{
Get mappability score for each bin in binned genome, and scale counts according to mappability score
}
