% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_counts.R
\name{rescaleXY}
\alias{rescaleXY}
\title{Rescale counts in chrX and chrY}
\usage{
rescaleXY(counts_df)
}
\arguments{
\item{counts_df}{dataframe with columns chr, start, end, and counts. Generally, the input to this function is the output of pbsR::rescaleMappability}
}
\description{
Rescale counts in chrX and chrY
}
